refactor: email 패키지 유틸 클래스 global로 이동
