--- src/main/java/com/gdschongik/gdsc/domain/email/constant/UnivEmailVerificationConstant.java
@@ -1,12 +1,0 @@
-package com.gdschongik.gdsc.domain.email.constant;
-
-public class UnivEmailVerificationConstant {
-
-    public static final String VERIFY_EMAIL_API_ENDPOINT = "/onboarding/verify-email?%s=";
-
-    public static final String VERIFY_EMAIL_REQUEST_PARAMETER_KEY = "token";
-
-    public static final String VERIFICATION_EMAIL_SUBJECT = "GDSC Hongik 이메일 인증 코드입니다.";
-
-    private UnivEmailVerificationConstant() {}
-}
