+++ src/main/java/com/gdschongik/gdsc/global/common/constant/EmailConstant.java
@@ -2,11 +2,12 @@ package com.gdschongik.gdsc.global.common.constant;
 
 public class EmailConstant {
 
-    public static final String SENDER_PERSONAL = "GDSC Hongik 운영팀";
-
-    public static final String SENDER_ADDRESS = "gdsc.hongik@gmail.com";
-
+    public static final String VERIFY_EMAIL_API_ENDPOINT = "/onboarding/verify-email?%s=";
+    public static final String VERIFY_EMAIL_REQUEST_PARAMETER_KEY = "token";
     public static final String HONGIK_UNIV_MAIL_DOMAIN = "@g.hongik.ac.kr";
+    public static final String SENDER_PERSONAL = "GDSC Hongik 운영팀";
+    public static final String SENDER_ADDRESS = "gdsc.hongik@gmail.com";
+    public static final String VERIFICATION_EMAIL_SUBJECT = "GDSC Hongik 이메일 인증 코드입니다.";
 
     private EmailConstant() {}
 }
